Implement data archive and loading. But you have to press home to save data, a direct kill will result in losing un saved data.

+++ Homepwner/BNRDetailViewController.m
@@ -69,7 +69,7 @@
     
     self.nameField.text = item.itemName;
     self.serialNumberField.text = item.serialNumber;
-    self.valueField.text = [NSString stringWithFormat:@"%d", item.valueInDollars];
+    self.valueField.text = [NSString stringWithFormat:@"%ld", (long)item.valueInDollars];
     
     static NSDateFormatter *dateFormatter = nil;
     if (!dateFormatter) {
